requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
websockets
paho-mqtt==2.0.0
greenlet==1.1.0
python-dotenv
starlette
requests
anyio